+++ Makefile
@@ -1,5 +1,5 @@
-BINARY := ipmi-catcher
-MAINMODULE := github.com/metal-stack/ipmi-catcher
+BINARY := bmc-catcher
+MAINMODULE := github.com/metal-stack/bmc-catcher
 COMMONDIR := $(or ${COMMONDIR},../builder)
 DOCKER_TAG := $(or ${GITHUB_TAG_NAME}, latest)
 
@@ -13,8 +13,8 @@ fmt:
 
 .PHONY: dockerimage
 dockerimage:
-	docker build -t metalstack/ipmi-catcher:${DOCKER_TAG} .
+	docker build -t metalstack/bmc-catcher:${DOCKER_TAG} .
 
 .PHONY: dockerpush
 dockerpush:
-	docker push metalstack/ipmi-catcher:${DOCKER_TAG}
+	docker push metalstack/bmc-catcher:${DOCKER_TAG}
