+++ docker-make.yaml
@@ -1,6 +1,6 @@
 ---
 version: '1'
-name: ipmi-catcher
+name: bmc-catcher
 username: metalstack
 registry-host: docker.io
 default-build-name: master
